requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pyyaml==6.0.1
python-multipart==0.0.6
watchdog==3.0.0